+++ src/extensions/inspector/props-config.js
@@ -1,3 +1,5 @@
+import { normalizeSource } from './utils.js';
+
 export const propsConfigView = {
     view: 'block',
     className: ['content', 'props-config'],
@@ -19,11 +21,44 @@ export const propsConfigView = {
             view: 'block',
             when: 'props != undefined',
             className: 'content-section props',
-            content: {
-                view: 'struct',
-                expanded: 2,
-                data: 'props'
-            }
+            content: [
+                {
+                    view: 'struct',
+                    expanded: 2,
+                    data: 'props'
+                },
+                {
+                    view: 'context',
+                    data: (data, context) =>
+                        context.host.view.getViewPropsTransition(data.props),
+                    whenData: true,
+                    content: [
+                        {
+                            view: 'block',
+                            className: 'transition-fn',
+                            data: data => ({
+                                ...data,
+                                query: typeof data.query === 'string' ? normalizeSource(data.query) : undefined
+                            }),
+                            content: {
+                                view: 'switch',
+                                content: [
+                                    { when: 'query is string', content: {
+                                        view: 'source',
+                                        data: '{ content: query, syntax: "jora", lineNum: false }'
+                                    } },
+                                    { content: 'struct:fn' }
+                                ]
+                            }
+                        },
+                        {
+                            view: 'struct',
+                            expanded: 2,
+                            data: 'props'
+                        }
+                    ]
+                }
+            ]
         },
         {
             view: 'block',