+++ src/views/text/text-numeric.js
@@ -1,11 +1,16 @@
 import { numDelim } from '../../core/utils/html.js';
 import usage from './text-numeric.usage.js';
 
+const props = `{
+    text: #.props has no 'text'?
+} | overrideProps()`;
+
 export default function(host) {
-    host.view.define('text-numeric', function (el, config, data) {
-        el.innerHTML = numDelim(data);
+    host.view.define('text-numeric', function (el, { text }) {
+        el.innerHTML = numDelim(text);
     }, {
         tag: 'span',
+        props,
         usage
     });
 }